# PERF: Cache hashing of categories