@fjetter 2021-03-03
Yes, there are `categoricals.Concat`

before
```
asv run --show-stderr --python=same --bench=categoricals.Concat       


· Running 6 total benchmarks (1 commits * 1 environments * 6 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_fjetter_miniconda3_envs_pandas-dev_bin_python
[  8.33%] ··· Running (categoricals.Concat.time_append_non_overlapping_index--)......
[ 58.33%] ··· categoricals.Concat.time_append_non_overlapping_index                  6.82±0.2ms
[ 66.67%] ··· categoricals.Concat.time_append_overlapping_index                      2.73±0.1ms
[ 75.00%] ··· categoricals.Concat.time_concat                                       2.66±0.05ms
[ 83.33%] ··· categoricals.Concat.time_concat_non_overlapping_index                  7.69±0.1ms
[ 91.67%] ··· categoricals.Concat.time_concat_overlapping_index                      3.34±0.2ms
[100.00%] ··· categoricals.Concat.time_union                                         2.90±0.1ms
```

and after

```
· Running 6 total benchmarks (1 commits * 1 environments * 6 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_fjetter_miniconda3_envs_pandas-dev_bin_python
[  8.33%] ··· Running (categoricals.Concat.time_append_non_overlapping_index--)......
[ 58.33%] ··· categoricals.Concat.time_append_non_overlapping_index                  1.35±0.3ms
[ 66.67%] ··· categoricals.Concat.time_append_overlapping_index                       78.7±30μs
[ 75.00%] ··· categoricals.Concat.time_concat                                       2.03±0.06ms
[ 83.33%] ··· categoricals.Concat.time_concat_non_overlapping_index                 2.03±0.05ms
[ 91.67%] ··· categoricals.Concat.time_concat_overlapping_index                        405±80μs
[100.00%] ··· categoricals.Concat.time_union                                         3.11±0.5ms
```

The improvements here (e.g. `categoricals.Concat.time_append_overlapping_index`) seem to be a bit too spectacular to be true. That's probably due to the caching since the benchmark iterations do not invalidate the cache. Do we have some experience with this kind of thing? I'll happily change the implementation if there are some recommendations to avoid this problem