@fjetter 2021-03-03
I believe the test failures are unrelated. I receive a lot of import errors regarding `html5lib`