@jreback 2021-03-03
also wouldn't mind the example you have above as an additional concat asv