@fjetter 2021-03-09
rebased and builds are green. Added a whatsnew entry but kept the logic for the caching as is (see comment above)

@jreback