@fjetter 2021-04-12
Reran both master and this branch


## master
```python
asv run --show-stderr --python=same --bench=categoricals.Concat
· Discovering benchmarks
· Running 6 total benchmarks (1 commits * 1 environments * 6 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_fjetter_mambaforge_envs_pandas-dev-39_bin_python
[  8.33%] ··· Running (categoricals.Concat.time_append_non_overlapping_index--)......
[ 58.33%] ··· categoricals.Concat.time_append_non_overlapping_index  4.23±0.08ms
[ 66.67%] ··· categoricals.Concat.time_append_overlapping_index         1.78±0ms
[ 75.00%] ··· categoricals.Concat.time_concat                            467±8μs
[ 83.33%] ··· categoricals.Concat.time_concat_non_overlapping_index  4.39±0.01ms
[ 91.67%] ··· categoricals.Concat.time_concat_overlapping_index      2.01±0.04ms
[100.00%] ··· categoricals.Concat.time_union                         1.68±0.08ms
```

## This branch
```python
· Discovering benchmarks
· Running 6 total benchmarks (1 commits * 1 environments * 6 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_fjetter_mambaforge_envs_pandas-dev-39_bin_python
[  8.33%] ··· Running (categoricals.Concat.time_append_non_overlapping_index--)......
[ 58.33%] ··· categoricals.Concat.time_append_non_overlapping_index     585±20μs
[ 66.67%] ··· categoricals.Concat.time_append_overlapping_index       21.3±0.6μs
[ 75.00%] ··· categoricals.Concat.time_concat                           371±20μs
[ 83.33%] ··· categoricals.Concat.time_concat_non_overlapping_index     886±60μs
[ 91.67%] ··· categoricals.Concat.time_concat_overlapping_index          162±1μs
[100.00%] ··· categoricals.Concat.time_union                         1.59±0.06ms
```

Note the drop in e.g. `time_concat_overlapping_index` is only that large because we do not invalidate the cache during benchmarks. realistic performance gains should be in the two digit percentages.

Using the above `make_df` functions with one-time execution gives something like a 20% drop

## master
```python
%time pd.concat(dfs)
CPU times: user 50.2 ms, sys: 2.08 ms, total: 52.3 ms
Wall time: 51.2 ms
```


## this branch
```python
%time pd.concat(dfs)
CPU times: user 36.2 ms, sys: 4.37 ms, total: 40.6 ms
Wall time: 39.4 ms
```